backend/requirements.txt
fastapi
uvicorn
transformers
torch
wandb
python-multipart
pydantic
requests
python-dotenv
